Fix mastodon feed disappearing when reopening that page.
This is not a correct way to fix the element but I couldnt figure out how to do it with pure react as the emfed library
does the rendering completely outside of react by replacing dom elements.

Current solution is to find out the element emfed is rendering into, store that to component and if user opens page second time, get the feed from that stored component.